# pkg/api/openapi.yaml
openapi: 3.0.3
info:
  title: Benchmarkoor API
  description: |
    API for the Benchmarkoor Ethereum execution client benchmarking tool.

    ## Authentication

    Most endpoints require authentication via either:
    - **Session cookie** (`benchmarkoor_session`) — obtained by logging in via `/auth/login` or GitHub OAuth
    - **Bearer token** — an API key prefixed with `bmk_`, passed in the `Authorization: Bearer <key>` header

    Endpoints under `/admin` additionally require the `admin` role.

    ## PostgREST-style Query Endpoints

    The `/index/query/*` endpoints support PostgREST-style filtering:
    - **Filtering:** Add query parameters in the form `column=operator.value` (e.g. `client=eq.geth`, `tests_total=gt.10`)
    - **Operators:** `eq`, `neq`, `gt`, `gte`, `lt`, `lte`, `like`, `in`, `is`
    - **Sorting:** `order=column.asc` or `order=column.desc` (comma-separated for multiple)
    - **Pagination:** `limit` and `offset`
    - **Column selection:** `select=col1,col2`
  version: 1.0.0
servers:
  - url: /api/v1
...
security: []
paths:
...
  /index/:
    get:
      operationId: listIndexedRuns
      tags: [index]
      summary: List all indexed runs
      description: |
        Returns all indexed runs across all discovery paths, sorted by
        timestamp descending. Same shape as `index.json` with an additional
        `discovery_path` field per entry.
      security:
        - cookieAuth: []
        - bearerAuth: []
        - {}
      responses:
        "200":
          description: Index listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndexResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /index/query/runs:
    get:
      operationId: queryRuns
      tags: [index]
      summary: Query indexed runs
      description: |
        Query indexed runs with PostgREST-style filtering, sorting, and pagination.

        **Filter columns:** id, discovery_path, run_id, timestamp, timestamp_end,
        suite_hash, status, termination_reason, has_result, instance_id, client,
        image, rollback_strategy, tests_total, tests_passed, tests_failed,
        indexed_at, reindexed_at, deletion_requested_at, deletion_error

        **Filter format:** `column=operator.value` (e.g. `client=eq.geth`)
      security:
        - cookieAuth: []
        - bearerAuth: []
        - {}
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 10000
          description: Maximum number of results
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Number of results to skip
        - name: select
          in: query
          schema:
            type: string
          description: Comma-separated list of columns to return
        - name: order
          in: query
          schema:
            type: string
          description: "Comma-separated sort directives (e.g. `timestamp.desc,client.asc`)"
      responses:
        "200":
          description: Query results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResultRuns"
        "400":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /index/query/test_stats:
    get:
      operationId: queryTestStats
      tags: [index]
      summary: Query test stat data
      description: |
        Query test stat data with PostgREST-style filtering, sorting, and pagination.

        **Filter columns:** id, suite_hash, test_name, run_id, client,
        total_gas_used, total_time_ns, total_mgas_s, setup_gas_used,
        setup_time_ns, setup_mgas_s, test_gas_used, test_time_ns,
        test_mgas_s, run_start, run_end
      security:
        - cookieAuth: []
        - bearerAuth: []
        - {}
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 10000
          description: Maximum number of results
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Number of results to skip
        - name: select
          in: query
          schema:
            type: string
          description: Comma-separated list of columns to return
        - name: order
          in: query
          schema:
            type: string
          description: "Comma-separated sort directives (e.g. `total_time_ns.desc`)"
      responses:
        "200":
          description: Query results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResultTestStats"
        "400":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /index/query/test_stats_block_logs:
    get:
      operationId: queryTestStatsBlockLogs
      tags: [index]
      summary: Query test stats block log data
      description: |
        Query per-block log data with PostgREST-style filtering, sorting, and pagination.
      security:
        - cookieAuth: []
        - bearerAuth: []
        - {}
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 10000
          description: Maximum number of results
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Number of results to skip
        - name: select
          in: query
          schema:
            type: string
          description: Comma-separated list of columns to return
        - name: order
          in: query
          schema:
            type: string
          description: "Comma-separated sort directives (e.g. `timing_total_ms.desc`)"
      responses:
        "200":
          description: Query results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResultTestStatsBlockLogs"
        "400":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Admin ───────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: benchmarkoor_session
    bearerAuth:
      type: http
      scheme: bearer
      description: API key prefixed with `bmk_`
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    IndexResponse:
      type: object
      properties:
        generated:
          type: integer
          format: int64
          description: Unix timestamp of when the index was generated
        entries:
          type: array
          items:
            $ref: "#/components/schemas/IndexEntry"
    IndexEntry:
      type: object
      properties:
        discovery_path:
          type: string
        run_id:
          type: string
        timestamp:
          type: integer
          format: int64
        timestamp_end:
          type: integer
          format: int64
        suite_hash:
          type: string
        status:
          type: string
        termination_reason:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: string
          description: Run metadata labels (key-value pairs)
        deletion_requested_at:
          type: integer
          format: int64
          description: |
            Unix timestamp (seconds) of when the run was queued for deletion.
            Omitted when the run is not queued.
        deletion_error:
          type: string
          description: Last deletion error for a queued run (omitted when empty)
        instance:
          type: object
          properties:
            id:
              type: string
            client:
              type: string
            image:
              type: string
            rollback_strategy:
              type: string
        tests:
          type: object
          properties:
            tests_total:
              type: integer
            tests_passed:
              type: integer
            tests_failed:
              type: integer
            steps:
              type: object
              properties:
                setup:
                  $ref: "#/components/schemas/StepStats"
                test:
                  $ref: "#/components/schemas/StepStats"
                cleanup:
                  $ref: "#/components/schemas/StepStats"
    StepStats:
      type: object
      properties:
        success:
          type: integer
        fail:
          type: integer
        duration:
          type: integer
          format: int64
        gas_used:
          type: integer
          format: uint64
        gas_used_duration:
          type: integer
          format: int64
        resource_totals:
          $ref: "#/components/schemas/ResourceTotals"
    ResourceTotals:
      type: object
      properties:
        cpu_usec:
          type: integer
          format: uint64
        memory_bytes:
          type: integer
          format: uint64
        memory_delta:
          type: integer
          format: int64
        disk_read_bytes:
          type: integer
          format: uint64
        disk_write_bytes:
          type: integer
          format: uint64
        disk_read_iops:
          type: integer
          format: uint64
        disk_write_iops:
          type: integer
          format: uint64
...
    QueryResultRuns:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/RunResponse"
        total:
          type: integer
          format: int64
        limit:
          type: integer
        offset:
          type: integer
    RunResponse:
      type: object
      properties:
        id:
          type: integer
        discovery_path:
          type: string
        run_id:
          type: string
        timestamp:
          type: integer
          format: int64
        timestamp_end:
          type: integer
          format: int64
        suite_hash:
          type: string
        status:
          type: string
        termination_reason:
          type: string
        has_result:
          type: boolean
        instance_id:
          type: string
        client:
          type: string
        image:
          type: string
        rollback_strategy:
          type: string
        tests_total:
          type: integer
        tests_passed:
          type: integer
        tests_failed:
          type: integer
        steps_json:
          type: object
          description: Raw step statistics JSON
        indexed_at:
          type: string
          format: date-time
        reindexed_at:
          type: string
          format: date-time
          nullable: true
        deletion_requested_at:
          type: string
          format: date-time
          nullable: true
          description: Set while the run is queued for deletion
        deletion_error:
          type: string
          description: Last deletion error for a queued run (omitted when empty)
    QueryResultTestStats:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/TestStatResponse"
        total:
          type: integer
          format: int64
        limit:
          type: integer
        offset:
          type: integer
    QueryResultTestStatsBlockLogs:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/TestStatsBlockLogResponse"
        total:
          type: integer
          format: int64
        limit:
          type: integer
        offset:
          type: integer
    TestStatResponse:
      type: object
      properties:
        id:
          type: integer
        suite_hash:
          type: string
        test_name:
          type: string
        run_id:
          type: string
        client:
          type: string
        total_gas_used:
          type: integer
          format: uint64
        total_time_ns:
          type: integer
          format: int64
        total_mgas_s:
          type: number
          format: double
        setup_gas_used:
          type: integer
          format: uint64
        setup_time_ns:
          type: integer
          format: int64
        setup_mgas_s:
          type: number
          format: double
        test_gas_used:
          type: integer
          format: uint64
        test_time_ns:
          type: integer
          format: int64
        test_mgas_s:
          type: number
          format: double
        run_start:
          type: integer
          format: int64
        run_end:
          type: integer
          format: int64
    TestStatsBlockLogResponse:
      type: object
      properties:
        id:
          type: integer
        suite_hash:
          type: string
        run_id:
          type: string
        test_name:
          type: string
        client:
          type: string
        block_number:
          type: integer
          format: uint64
        block_hash:
          type: string
        block_gas_used:
          type: integer
          format: uint64
        block_tx_count:
          type: integer
        timing_execution_ms:
          type: number
          format: double
        timing_state_read_ms:
          type: number
          format: double
        timing_state_hash_ms:
          type: number
          format: double
        timing_commit_ms:
          type: number
          format: double
        timing_total_ms:
          type: number
          format: double
        throughput_mgas_per_sec:
          type: number
          format: double